refactor(Pipelines): Split schema.py in types, queries, mutations for clarity
