+++ hexa/pipelines/schema/__init__.py
@@ -1,0 +1,14 @@
+import pathlib
+
+from ariadne import load_schema_from_path
+
+from .mutations import bindables as mutations_bindables
+from .queries import bindables as queries_bindables
+from .types import bindables as types_bindables
+
+pipelines_type_defs = load_schema_from_path(
+    f"{pathlib.Path(__file__).parent.parent.resolve()}/graphql/schema.graphql"
+)
+
+
+pipelines_bindables = types_bindables + queries_bindables + mutations_bindables
