+++ hexa/pipelines/schema/mutations.py
@@ -1,30 +1,16 @@
 import base64
-import pathlib
 
-from ariadne import (
-    EnumType,
-    MutationType,
-    ObjectType,
-    QueryType,
-    UnionType,
-    convert_kwargs_to_snake_case,
-    load_schema_from_path,
-)
+from ariadne import MutationType, convert_kwargs_to_snake_case
 from django.conf import settings
 from django.core.exceptions import PermissionDenied
 from django.db import IntegrityError
 from django.http import HttpRequest
-from django.urls import reverse
-from sentry_sdk import capture_exception
+from google.api_core.exceptions import NotFound
 
-from hexa.core.graphql import result_page
 from hexa.databases.utils import get_table_definition
-from hexa.files.api import NotFound, get_storage
-from hexa.workspaces.models import Workspace, WorkspaceMembershipRole
-from hexa.workspaces.schema.types import workspace_permissions
-
-from .authentication import PipelineRunUser
-from .models import (
+from hexa.files.api import get_storage
+from hexa.pipelines.authentication import PipelineRunUser
+from hexa.pipelines.models import (
     InvalidTimeoutValueError,
     Pipeline,
     PipelineDoesNotSupportParametersError,
@@ -33,300 +19,14 @@ from .models import (
     PipelineRunTrigger,
     PipelineVersion,
 )
+from hexa.workspaces.models import Workspace, WorkspaceMembershipRole
 
-pipelines_type_defs = load_schema_from_path(
-    f"{pathlib.Path(__file__).parent.resolve()}/graphql/schema.graphql"
-)
+pipelines_mutations = MutationType()
 
 
 # ease mocking
 def get_bucket_object(bucket_name, file):
     return get_storage().get_bucket_object(bucket_name, file)
-
-
-@workspace_permissions.field("createPipeline")
-def resolve_workspace_permissions_create_pipeline(obj: Workspace, info, **kwargs):
-    request = info.context["request"]
-
-    return request.user.is_authenticated and request.user.has_perm(
-        "pipelines.create_pipeline", obj
-    )
-
-
-pipeline_permissions = ObjectType("PipelinePermissions")
-pipeline_parameter = ObjectType("PipelineParameter")
-pipeline_run_status_enum = EnumType("PipelineRunStatus", PipelineRun.STATUS_MAPPINGS)
-pipeline_run_order_by_enum = EnumType(
-    "PipelineRunOrderBy",
-    {
-        "EXECUTION_DATE_DESC": "-execution_date",
-        "EXECUTION_DATE_ASC": "execution_date",
-    },
-)
-pipeline_object = ObjectType("Pipeline")
-generic_output_object = ObjectType("GenericOutput")
-
-pipeline_run_output_union = UnionType("PipelineRunOutput")
-
-
-@pipeline_run_output_union.type_resolver
-def resolve_run_output_type(obj, *_):
-    if "columns" in obj:
-        return "DatabaseTable"
-    elif obj["type"] == "file" or obj["type"] == "directory":
-        return "BucketObject"
-
-    return "GenericOutput"
-
-
-@pipeline_parameter.field("name")
-def resolve_pipeline_parameter_code(parameter, info, **kwargs):
-    name = parameter.get("name")
-    if name is None:
-        name = parameter["code"]
-    return name
-
-
-@pipeline_parameter.field("required")
-def resolve_pipeline_parameter_required(parameter, info, **kwargs):
-    return parameter.get("required", False)
-
-
-@pipeline_parameter.field("multiple")
-def resolve_pipeline_parameter_multiple(parameter, info, **kwargs):
-    return parameter.get("multiple", False)
-
-
-@pipeline_permissions.field("update")
-def resolve_pipeline_permissions_update(pipeline: Pipeline, info, **kwargs):
-    request = info.context["request"]
-    return request.user.is_authenticated and request.user.has_perm(
-        "pipelines.update_pipeline", pipeline
-    )
-
-
-@pipeline_permissions.field("delete")
-def resolve_pipeline_permissions_delete(pipeline: Pipeline, info, **kwargs):
-    request = info.context["request"]
-    return request.user.is_authenticated and request.user.has_perm(
-        "pipelines.delete_pipeline", pipeline
-    )
-
-
-@pipeline_permissions.field("run")
-def resolve_pipeline_permissions_run(pipeline: Pipeline, info, **kwargs):
-    request = info.context["request"]
-    return request.user.is_authenticated and request.user.has_perm(
-        "pipelines.run_pipeline", pipeline
-    )
-
-
-@pipeline_permissions.field("deleteVersion")
-def resolve_pipeline_permissions_delete_version(pipeline: Pipeline, info, **kwargs):
-    request = info.context["request"]
-    return request.user.is_authenticated and request.user.has_perm(
-        "pipelines.delete_pipeline_version", pipeline
-    )
-
-
-@pipeline_permissions.field("schedule")
-def resolve_pipeline_permissions_schedule(pipeline: Pipeline, info, **kwargs):
-    request = info.context["request"]
-    return (
-        request.user.is_authenticated
-        and request.user.has_perm("pipelines.run_pipeline", pipeline)
-        and pipeline.last_version.is_schedulable
-    )
-
-
-@pipeline_object.field("webhookUrl")
-def resolve_pipeline_webhook_url(pipeline: Pipeline, info):
-    request = info.context["request"]
-    if pipeline.webhook_enabled:
-        return request.build_absolute_uri(reverse("pipelines:run", args=[pipeline.id]))
-
-
-@pipeline_object.field("currentVersion")
-def resolve_pipeline_current_version(pipeline: Pipeline, info, **kwargs):
-    return pipeline.last_version
-
-
-@pipeline_object.field("permissions")
-def resolve_pipeline_permissions(pipeline: Pipeline, info, **kwargs):
-    return pipeline
-
-
-@pipeline_object.field("versions")
-def resolve_pipeline_versions(pipeline: Pipeline, info, **kwargs):
-    qs = pipeline.versions.all()
-    return result_page(
-        queryset=qs, page=kwargs.get("page", 1), per_page=kwargs.get("perPage")
-    )
-
-
-@pipeline_object.field("runs")
-def resolve_pipeline_runs(pipeline: Pipeline, info, **kwargs):
-    qs = PipelineRun.objects.filter(pipeline=pipeline)
-
-    order_by = kwargs.get("orderBy", None)
-    if order_by is not None:
-        qs = qs.order_by(order_by)
-    else:
-        qs = qs.order_by("-execution_date")
-
-    return result_page(
-        queryset=qs, page=kwargs.get("page", 1), per_page=kwargs.get("perPage")
-    )
-
-
-@pipeline_object.field("recipients")
-def resolve_pipeline_recipients(pipeline: Pipeline, info, **kwargs):
-    return pipeline.pipelinerecipient_set.all()
-
-
-pipeline_run_object = ObjectType("PipelineRun")
-
-
-@pipeline_run_object.field("triggerMode")
-def resolve_pipeline_run_trigger_mode(run: PipelineRun, info, **kwargs):
-    return run.trigger_mode
-
-
-@pipeline_run_object.field("duration")
-def resolve_pipeline_run_duration(run: PipelineRun, info, **kwargs):
-    return int(run.duration.total_seconds()) if run.duration is not None else 0
-
-
-@pipeline_run_object.field("config")
-def resolve_pipeline_run_config(run: PipelineRun, info, **kwargs):
-    return run.config
-
-
-@pipeline_run_object.field("code")
-def resolve_pipeline_run_code(run: PipelineRun, info, **kwargs):
-    return base64.b64encode(run.get_code()).decode("ascii")
-
-
-pipeline_run_object.set_alias("progress", "current_progress")
-pipeline_run_object.set_alias("logs", "run_logs")
-pipeline_run_object.set_alias("version", "pipeline_version")
-
-
-pipeline_version_object = ObjectType("PipelineVersion")
-
-
-@pipeline_version_object.field("zipfile")
-def resolve_pipeline_version_zipfile(version: PipelineVersion, info, **kwargs):
-    return base64.b64encode(version.zipfile).decode("ascii")
-
-
-pipelines_query = QueryType()
-
-
-@pipelines_query.field("pipelines")
-def resolve_pipelines(_, info, **kwargs):
-    request: HttpRequest = info.context["request"]
-    if kwargs.get("workspaceSlug", None):
-        try:
-            ws = Workspace.objects.filter_for_user(request.user).get(
-                slug=kwargs.get("workspaceSlug")
-            )
-            qs = (
-                Pipeline.objects.filter_for_user(request.user)
-                .filter(workspace=ws)
-                .order_by("name", "id")
-            )
-        except Workspace.DoesNotExist:
-            qs = Pipeline.objects.none()
-    else:
-        qs = Pipeline.objects.filter_for_user(request.user).order_by("name", "id")
-
-    return result_page(
-        queryset=qs, page=kwargs.get("page", 1), per_page=kwargs.get("perPage")
-    )
-
-
-@pipelines_query.field("pipeline")
-def resolve_pipeline(_, info, **kwargs):
-    request: HttpRequest = info.context["request"]
-    try:
-        return Pipeline.objects.filter_for_user(request.user).get(id=kwargs["id"])
-    except Pipeline.DoesNotExist:
-        return None
-
-
-@pipelines_query.field("pipelineByCode")
-def resolve_pipeline_by_code(_, info, **kwargs):
-    request: HttpRequest = info.context["request"]
-    try:
-        pipeline = Pipeline.objects.filter_for_user(request.user).get(
-            workspace__slug=kwargs["workspaceSlug"], code=kwargs["code"]
-        )
-    except Pipeline.DoesNotExist:
-        pipeline = None
-
-    return pipeline
-
-
-@pipelines_query.field("pipelineRun")
-def resolve_pipeline_run(_, info, **kwargs):
-    request: HttpRequest = info.context["request"]
-
-    if not request.user.is_authenticated:
-        return None
-
-    run_id = kwargs["id"]
-    try:
-        if isinstance(request.user, PipelineRunUser):
-            qs = PipelineRun.objects.filter(id=request.user.pipeline_run.id).exclude(
-                state__in=[PipelineRunState.SUCCESS, PipelineRunState.FAILED]
-            )
-        else:
-            qs = PipelineRun.objects.filter_for_user(request.user)
-
-        return qs.get(id=run_id)
-
-    except PipelineRun.DoesNotExist:
-        return None
-
-
-@pipeline_run_object.field("outputs")
-def resolve_pipeline_run_outputs(run: PipelineRun, info, **kwargs):
-    result = []
-    workspace = run.pipeline.workspace
-    for output in run.outputs:
-        try:
-            if output["type"] == "file":
-                result.append(
-                    get_bucket_object(
-                        workspace.bucket_name,
-                        output["uri"][len(f"gs://{workspace.bucket_name}/") :],
-                    )
-                )
-            elif output["type"] == "db":
-                table_data = get_table_definition(workspace, output["name"])
-                if table_data:
-                    result.append(table_data)
-                else:
-                    raise Exception(
-                        f"Table {output['name']} not found or connection error"
-                    )
-            else:
-                result.append(output)
-        except Exception as e:
-            # Table or Bucket object might be deleted
-            capture_exception(e)
-            continue
-
-    return result
-
-
-@pipeline_run_object.field("datasetVersions")
-def resolve_pipeline_run_dataset_version(run: PipelineRun, info, **kwargs):
-    return run.dataset_versions.all()
-
-
-pipelines_mutations = MutationType()
 
 
 @pipelines_mutations.field("createPipeline")
@@ -640,16 +340,4 @@ def resolve_add_pipeline_output(_, info, **kwargs):
     return {"success": True, "errors": []}
 
 
-pipelines_bindables = [
-    pipelines_query,
-    pipelines_mutations,
-    pipeline_object,
-    pipeline_parameter,
-    pipeline_run_object,
-    pipeline_run_status_enum,
-    pipeline_run_order_by_enum,
-    pipeline_version_object,
-    pipeline_permissions,
-    pipeline_run_output_union,
-    generic_output_object,
-]
+bindables = [pipelines_mutations]
